+++ CHANGELOG.md
@@ -9,7 +9,13 @@ The `Unreleased` section name is replaced by the expected version of next releas
 ## [Unreleased]
 
 ### Added
+
+- `Kafka.Producer`: Expose `acks`, `linger`, `compression` options [#78](https://github.com/jet/propulsion/pull/78)
+
 ### Changed
+
+- `Kafka`: Adjust default linger to 5ms [#78](https://github.com/jet/propulsion/pull/78)
+
 ### Removed
 ### Fixed